Trekk ut hjemler i egen fil
